Completely rewrite spub-git-delta-navigate
The previous design had a flaw which stemmed from how stdin was being
read into the script. The flaw was that the script required to read the
entirety of stdin BEFORE it could pipe the procesed lines into less.

This made any other script which uses this, e.g. logvo, slow, more so
when stdin is very large.

The new spub-git-delta-navigate avoids reading all stdin by constantly
printing to stdout in a while loop.